forex_bot/requirements.txt
alpaca-py>=0.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
keras>=2.14.0
tensorflow>=2.14.0
joblib>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
